完善根据id删除分类的业务需求，补充了当该分类关联了菜品/套餐的时候无法进行删除的业务逻辑

+++ sky-backend/sky-server/src/main/java/com/sky/controller/admin/CategoryController.java
@@ -66,7 +66,7 @@ public class CategoryController {
      */
     @DeleteMapping()
     @ApiOperation("根据id删除分类")
-    public Result deleteCategoryById(@RequestParam Integer id) {
+    public Result deleteCategoryById(@RequestParam Long id) {
         categoryService.deleteCategoryById(id);
         return Result.success();
     }
